requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
plotly==5.17.0
dash==2.14.2
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
scipy==1.11.4
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
twilio==8.10.0
pytest==7.4.3
pytest-asyncio==0.21.1
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
aiofiles==23.2.1
websockets==12.0
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
jinja2==3.1.2